move integration tests to separate directory
- Name cmdtests makes sense to me because we test commands
- Tests files use package name cmdtests_test instead of cmdtests:
  - goimports at least on my machine tries to import cmdtests even when
  then name of the package is the same which causes cycle imports. I'm
  not sure why it happens, most probably because of build tags.
  - in my opinion both cmdtests_test and cmdtests make sense so I just
  kept it like this
