+++ cmdtests/parse_test.go
@@ -1,6 +1,6 @@
 // +build integration
 
-package cmd
+package cmdtests_test
 
 import (
 	"bytes"
@@ -12,14 +12,13 @@ import (
 	"strings"
 	"testing"
 
-	cmdtest "github.com/src-d/engine/cmd/test-utils"
-
+	"github.com/src-d/engine/cmdtests"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 )
 
 type ParseTestSuite struct {
-	cmdtest.IntegrationSuite
+	cmdtests.IntegrationSuite
 	testDir string
 }
