+++ cmdtests/testdata/LICENSE
