+++ cmdtests/testdata/hello-py3.py
