+++ cmdtests/testdata/hello.bash
