+++ cmdtests/testdata/hello.cpp
