+++ cmdtests/testdata/hello.cs
