+++ cmdtests/testdata/hello.go
