+++ cmdtests/testdata/hello.java
