+++ cmdtests/testdata/hello.js
