+++ cmdtests/testdata/hello.php
