+++ cmdtests/testdata/hello.py
