+++ cmdtests/testdata/hello.rb
